pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "djinntoux"
version = "0.0.4.4"
description = "Generic reusable bits for Django projects"
readme = "README.md"
requires-python = ">= 3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",       # for scraping
    "requests>=2.32.3",             # for scraping, also installs urllib3
    "uuid6>=2024.7.10",             # for abstract models

    "django>=5.2a1",                # standard dep
    "django-environ>=0.12.0",       # standard dep
    "django-htmlmin>=0.11.0",       # standard dep
    "django-tree-queries>=0.19.0",  # standard dep
    "djangoql>=0.18.1",             # standard dep
    "gunicorn>=23.0.0",             # standard dep
    "psycopg>=3.2.4",               # standard dep

    "django-timezone-field>=7.1",   # for user model but no autocomplete

    "pip-autoremove>=0.10.0"        # extra
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Repository = "https://github.com/venteto/djinntoux"